Rearrangement of every search feature in SearchForm.js
