+++ package.json
@@ -11,7 +11,8 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.2.0",
     "react-swipeable-views": "^0.13.3",
-    "styled-components": "^5.0.0"
+    "styled-components": "^5.0.0",
+    "yup": "^0.28.1"
   },
   "scripts": {
     "start": "react-scripts start",
